Stats and persistence
Created the 4 stat fields for hens (love, size, fitness, feather quality). Created a system to save the attributes of each hen (stored as a JSON structure within playerPrefs), and a spawning system that spawns hens at the start of the Raise scene based on the saved data (or spawns 4 default hens if there is no save).